scripts/requirements.txt
chardet==3.0.4
dbus-python==1.2.4
decorator==4.2.1
ethtool==0.14
gpg==1.13.1
idna==2.5
iniparse==0.4
libcomps==0.1.18
pygobject==3.28.3
pyinotify==0.9.6
PySocks==1.6.8
python-dateutil==2.6.1
PyYAML==6.0.1
requests==2.20.0
rpm==4.14.3
six==1.11.0
subscription-manager==1.28.40
syspurpose==1.28.40
systemd-python==234
urllib3==1.24.2